+++ db/structure.sql
@@ -1348,7 +1348,7 @@ CREATE OR REPLACE VIEW public.student_lesson_summaries AS
            FROM ((((public.students s
              JOIN public.groups g ON ((g.id = s.group_id)))
              JOIN public.lessons l ON ((g.id = l.group_id)))
-             LEFT JOIN public.grades ON (((grades.student_id = s.id) AND (grades.lesson_id = l.id))))
+             LEFT JOIN public.grades ON (((grades.student_id = s.id) AND (grades.lesson_id = l.id) AND (grades.deleted_at IS NULL))))
              LEFT JOIN public.absences a ON (((a.student_id = s.id) AND (a.lesson_id = l.id))))
           GROUP BY s.id, l.id, a.id
         UNION
@@ -1366,7 +1366,7 @@ CREATE OR REPLACE VIEW public.student_lesson_summaries AS
                 END AS absent
            FROM ((((public.lessons l
              JOIN public.groups g ON ((g.id = l.group_id)))
-             JOIN public.grades ON ((grades.lesson_id = l.id)))
+             JOIN public.grades ON (((grades.lesson_id = l.id) AND (grades.deleted_at IS NULL))))
              JOIN public.students s ON ((grades.student_id = s.id)))
              LEFT JOIN public.absences a ON (((a.student_id = s.id) AND (a.lesson_id = l.id))))
           GROUP BY s.id, l.id, a.id) united;
@@ -1386,8 +1386,7 @@ CREATE OR REPLACE VIEW public.lesson_skill_summaries AS
      JOIN public.subjects su ON ((su.id = l.subject_id)))
      JOIN public.assignments a ON ((su.id = a.subject_id)))
      JOIN public.skills sk ON ((a.skill_id = sk.id)))
-     LEFT JOIN public.grades g ON (((g.lesson_uid = l.uid) AND (g.skill_id = sk.id))))
-  WHERE (g.deleted_at IS NULL)
+     LEFT JOIN public.grades g ON (((g.lesson_uid = l.uid) AND (g.skill_id = sk.id) AND (g.deleted_at IS NULL))))
   GROUP BY l.uid, sk.id;
 
 
@@ -1656,6 +1655,8 @@ INSERT INTO "schema_migrations" (version) VALUES
 ('20191102173151'),
 ('20191102200044'),
 ('20191102234931'),
-('20191103021012');
+('20191103021012'),
+('20191107004248'),
+('20191107010120');